Add support for auto backup preconditions
In certain cases it doesn't make sense to proceed with an automated
backup unless a precondition is met. A great example is a storage server
that doesn't have its storage file system mounted. Running a backup in
this case would result in an empty archive (as the mountpoint would be
an empty directory) and potentially increase the risk of data loss.